1/requirements2.txt
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
customtkinter==5.2.1
darkdetect==0.8.0
dict2xml==1.7.3
idna==3.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==9.5.0
platformdirs==3.11.0
Pygments==2.16.1
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
tkhtmlview==0.2.0
urllib3==2.0.7
xmltodict==0.13.0
